% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here-funs.R
\name{here_ref}
\alias{here_ref}
\alias{here}
\alias{here_app}
\title{Place thesis components.}
\usage{
here_ref()

here_app()
}
\description{
Functions for placing dissertation components. These functions write LaTeX and HTML code behind-the-scenes so you don't have to.
}
\examples{
\dontrun{
here_ref()
here_app()
}

}
