% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chapter.R
\name{new_chapter}
\alias{new_chapter}
\title{Create New Chapter}
\usage{
new_chapter(name = NULL)
}
\arguments{
\item{name}{Name of file.}
}
\description{
Create a new chapter in your thesis using a R Markdown file
}
\examples{

\dontrun{
new_chapter('sweet-chapter')
}

}
