% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-tinytex-funs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{install_tinytex}
\alias{uninstall_tinytex}
\alias{tl_pkgs}
\alias{pdflatex}
\alias{lualatex}
\alias{xelatex}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tinytex}{\code{\link[tinytex]{install_tinytex}}, \code{\link[tinytex:latexmk]{lualatex}}, \code{\link[tinytex:latexmk]{pdflatex}}, \code{\link[tinytex]{tl_pkgs}}, \code{\link[tinytex:install_tinytex]{uninstall_tinytex}}, \code{\link[tinytex:latexmk]{xelatex}}}
}}

