% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torontodown.R
\docType{package}
\name{torontodown}
\alias{torontodown}
\alias{torontodown-package}
\title{torontodown: A package for creating Masters and Doctoral theses
using R Markdown}
\description{
This package creates theses/dissertations for University of Toronto.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

