% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_opts.R
\name{yaml_opts}
\alias{yaml_opts}
\title{YAML Key-Value Options}
\usage{
yaml_opts()
}
\description{
Prints a list of the options specific to the University of Toronto thesis class that are passed to the TeX template specified as document class options.
}
